+++ src/app/main.cpp
@@ -1,4 +1,5 @@
 #include <QApplication>
+#include <QIcon>
 
 #include "ui/MainWindow.h"
 
@@ -7,7 +8,8 @@ int main(int argc, char* argv[])
     QApplication app(argc, argv);
     app.setOrganizationName("AstroTracker");
     app.setApplicationName("AstroTracker");
-    app.setApplicationVersion("0.1.0");
+    app.setApplicationVersion(QStringLiteral(ASTROTRACKER_VERSION));
+    app.setWindowIcon(QIcon(QStringLiteral(":/favicon/app-512.png")));
 
     MainWindow window;
     window.show();